==> code/00_Count.C4.Rmd <==
---
title: "00_Count.C4"
output: html_document
---

## Install packages #
```{r}
rm(list = ls())
options(stringsAsFactors = FALSE)

library(tidyverse)
```

## Import C4 imputation results #
```{r}
c4.imputed <- read.delim("../data/C4.imputed.multethnic.transposed.txt", header = FALSE, na.strings = c("", "NA"))  

c4.imputed <- c4.imputed %>% rename(subjectkey = V1, allele1 = V2, allele2 = V3)
colnames(c4.imputed)[4:ncol(c4.imputed)] <- paste0("structure", 1:30)
```

## Calculate number of copies #
```{r}
c4.imputed$c4a_cn1 <- as.numeric(ifelse(
  c4.imputed$allele1 %in% c("1", "2", "5", "6"), 0, 
    ifelse(c4.imputed$allele1 %in% c("3", "4", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "24", "25", "26"), 1, 
         ifelse(c4.imputed$allele1 %in% c("19", "20", "21", "22", "23", "27", "28", "29"), 2, "UNK")))
)

c4.imputed$c4b_cn1 <- as.numeric(ifelse(
  c4.imputed$allele1 %in% c("3", "4", "19", "20", "21", "22", "23"), 0, 
    ifelse(c4.imputed$allele1 %in% c("1", "2", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "27", "28", "29"), 1, 
         ifelse(c4.imputed$allele1 %in% c("5", "6", "24", "25", "26"), 2, "UNK")))
)

c4.imputed$herv_cn1 <- as.numeric(ifelse(
  c4.imputed$allele1 %in% c("1", "3"), 0, 
    ifelse(c4.imputed$allele1 %in% c("2", "4", "5", "7", "8", "9", "10", "11", "12", "13", "14", "15", "19", "24", "27"), 1, 
         ifelse(c4.imputed$allele1 %in% c("6", "16", "17", "18", "20", "21", "22", "23", "25", "28"), 2,
                ifelse(c4.imputed$allele1 %in% c("26", "29"), 3, "UNK"))))
)

c4.imputed$c4a_cn2 <- as.numeric(ifelse(
  c4.imputed$allele2 %in% c("1", "2", "5", "6"), 0, 
    ifelse(c4.imputed$allele2 %in% c("3", "4", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "24", "25", "26"), 1, 
         ifelse(c4.imputed$allele2 %in% c("19", "20", "21", "22", "23", "27", "28", "29"), 2, "UNK")))
)

c4.imputed$c4b_cn2 <- as.numeric(ifelse(
  c4.imputed$allele2 %in% c("3", "4", "19", "20", "21", "22", "23"), 0, 
    ifelse(c4.imputed$allele2 %in% c("1", "2", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "27", "28", "29"), 1, 
         ifelse(c4.imputed$allele2 %in% c("5", "6", "24", "25", "26"), 2, "UNK")))
)

c4.imputed$herv_cn2 <- as.numeric(ifelse(
  c4.imputed$allele2 %in% c("1", "3"), 0, 
    ifelse(c4.imputed$allele2 %in% c("2", "4", "5", "7", "8", "9", "10", "11", "12", "13", "14", "15", "19", "27", "24"), 1, 
         ifelse(c4.imputed$allele2 %in% c("6", "16", "17", "18", "20", "21", "22", "23", "25", "28"), 2,
                ifelse(c4.imputed$allele2 %in% c("26", "29"), 3, "UNK"))))
)
```

## Calculate total number of c4a, c4b, and C4-HERV elements #
```{r}
c4.imputed$c4a_total <- as.numeric(c4.imputed$c4a_cn1 + c4.imputed$c4a_cn2)
c4.imputed$c4b_total <- as.numeric(c4.imputed$c4b_cn1 + c4.imputed$c4b_cn2)
c4.imputed$c4herv_total <- as.numeric(c4.imputed$herv_cn1 + c4.imputed$herv_cn2)
```

## Calculate overall imputation accuracy #
```{r}
c4.imputed$avg_dosage <- NA
for (i in 1:nrow(c4.imputed)) {
  c4.imputed$avg_dosage[i] <- 0.5 * (c4.imputed[i, paste0("structure", c4.imputed$allele1[i])] +
                                       c4.imputed[i, paste0("structure", c4.imputed$allele2[i])])
}

idx <- which(c4.imputed$allele1 == c4.imputed$allele2)
c4.imputed$avg_dosage[idx] <- c4.imputed$avg_dosage[idx] / 2

hist(c4.imputed$avg_dosage, breaks = 100)
```

## Remove individuals with low quality imputation #
```{r}
c4.imputed <- c4.imputed %>% filter(avg_dosage > 0.7) 
```

## Calculate probabilistic dosage #
```{r}
prob_mat <- as.matrix(c4.imputed %>% select(starts_with("structure")))
dim(prob_mat)

c4a_vector <- c(0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, 2, 1, 1, 1, 2, 2, 2, 0)
c4a_dosage <- prob_mat %*% c4a_vector
plot(c4a_dosage, c4.imputed$c4a_total)
cor(c4a_dosage, c4.imputed$c4a_total)  

c4b_vector <- c(1, 1, 0, 0, 2, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 2, 2, 2, 1, 1, 1, 0)
c4b_dosage <- prob_mat %*% c4b_vector
plot(c4b_dosage, c4.imputed$c4b_total)
cor(c4b_dosage, c4.imputed$c4b_total)  

c4.imputed$c4a_dosage <- c4a_dosage
c4.imputed$c4b_dosage <- c4b_dosage
```

## Assign structural forms #
```{r}
c4.imputed$c4_allele1_structure <- as.factor(c4.imputed$allele1)
c4.imputed$c4_allele2_structure <- as.factor(c4.imputed$allele2)

levels(c4.imputed$c4_allele1_structure)
levels(c4.imputed$c4_allele2_structure)

which(colnames(c4.imputed) == "c4_allele1_structure" ) 
which(colnames(c4.imputed) == "c4_allele2_structure" ) 

for (i in 46:47) {
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="1"] <- "BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="2"] <- "BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="3"] <- "AS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="4"] <- "AL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="5"] <- "BL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="6"] <- "BL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="7"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="8"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="9"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="10"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="11"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="12"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="13"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="14"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="15"] <- "AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="16"] <- "AL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="17"] <- "AL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="18"] <- "AL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="19"] <- "AL-AS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="20"] <- "AL-AL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="21"] <- "AL-AL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="22"] <- "AL-AL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="23"] <- "AL-AL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="24"] <- "AL-BS-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="25"] <- "AL-BL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="26"] <- "AL-BL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="27"] <- "AL-AS-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="28"] <- "AL-AL-BS"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="29"] <- "AL-AL-BL"
levels(c4.imputed[, i])[levels(c4.imputed[, i])=="30"] <- "UNK"
}

levels(c4.imputed$c4_allele1_structure)
levels(c4.imputed$c4_allele2_structure)
```

## Assign c4a/B expression; c4a expression = (0.47*AL) + (0.47*AS) + (0.20*BL); c4b expression = (1.03*BL) + (0.88*BS) #
```{r}
# c4a expression
c4.imputed$c4a_allele1_expression <- as.numeric(ifelse(
  c4.imputed$allele1 %in% c("1"), 0, 
    ifelse(c4.imputed$allele1 %in% c("2", "5"), 0.2, 
         ifelse(c4.imputed$allele1 %in% c("3", "4", "7", "8", "9", "10", "11", "12", "13", "14", "15", "24"), 0.47, 
                ifelse(c4.imputed$allele1 %in% c("6"), 0.4, 
                      ifelse(c4.imputed$allele1 %in% c("16", "17", "18", "25"), 0.67, 
                             ifelse(c4.imputed$allele1 %in% c("19", "20", "21", "22", "23", "27", "28"), 0.94,
                                    ifelse(c4.imputed$allele1 %in% c("26"), 0.87, 
                                           ifelse(c4.imputed$allele1 %in% c("29"), 1.14, "UNK"))))))))
)

c4.imputed$c4a_allele2_expression <- as.numeric(ifelse(
  c4.imputed$allele2 %in% c("1"), 0, 
    ifelse(c4.imputed$allele2 %in% c("2", "5"), 0.2, 
         ifelse(c4.imputed$allele2 %in% c("3", "4", "7", "8", "9", "10", "11", "12", "13", "14", "15", "24"), 0.47, 
                ifelse(c4.imputed$allele2 %in% c("6"), 0.4, 
                      ifelse(c4.imputed$allele2 %in% c("16", "17", "18", "25"), 0.67, 
                             ifelse(c4.imputed$allele2 %in% c("19", "20", "21", "22", "23", "27", "28"), 0.94,
                                    ifelse(c4.imputed$allele2 %in% c("26"), 0.87, 
                                           ifelse(c4.imputed$allele2 %in% c("29"), 1.14, "UNK"))))))))
)

c4.imputed$c4a_expression <- (c4.imputed$c4a_allele1_expression + c4.imputed$c4a_allele2_expression)
range(c4.imputed$c4a_expression)
hist(c4.imputed$c4a_expression)
plot(c4.imputed$c4a_total, c4.imputed$c4a_expression)
cor(c4.imputed$c4a_total, c4.imputed$c4a_expression)
plot(c4.imputed$c4a_dosage, c4.imputed$c4a_expression)
cor(c4.imputed$c4a_dosage, c4.imputed$c4a_expression)

# c4b expression
c4.imputed$c4b_allele1_expression <- as.numeric(ifelse(
  c4.imputed$allele1 %in% c("1", "7", "8", "9", "10", "11", "12", "13", "14", "15", "27", "28"), 0.88, 
    ifelse(c4.imputed$allele1 %in% c("2", "16", "17", "18", "29"), 1.03, 
         ifelse(c4.imputed$allele1 %in% c("3", "4", "19", "20", "21", "22", "23"), 0, 
                ifelse(c4.imputed$allele1 %in% c("5", "25"), 1.91, 
                      ifelse(c4.imputed$allele1 %in% c("6", "26"), 2.06, 
                             ifelse(c4.imputed$allele1 %in% c("24"), 1.76, "UNK"))))))
)

c4.imputed$c4b_allele2_expression <- as.numeric(ifelse(
  c4.imputed$allele2 %in% c("1", "7", "8", "9", "10", "11", "12", "13", "14", "15", "27", "28"), 0.88, 
    ifelse(c4.imputed$allele2 %in% c("2", "16", "17", "18", "29"), 1.03, 
         ifelse(c4.imputed$allele2 %in% c("3", "4", "19", "20", "21", "22", "23"), 0, 
                ifelse(c4.imputed$allele2 %in% c("5", "25"), 1.91, 
                      ifelse(c4.imputed$allele2 %in% c("6", "26"), 2.06, 
                             ifelse(c4.imputed$allele2 %in% c("24"), 1.76, "UNK"))))))
)

c4.imputed$c4b_expression <- (c4.imputed$c4b_allele1_expression + c4.imputed$c4b_allele2_expression)
range(c4.imputed$c4b_expression)
hist(c4.imputed$c4b_expression)
plot(c4.imputed$c4b_total, c4.imputed$c4b_expression)
cor(c4.imputed$c4b_total, c4.imputed$c4b_expression)
plot(c4.imputed$c4b_dosage, c4.imputed$c4b_expression)
cor(c4.imputed$c4b_dosage, c4.imputed$c4b_expression)

plot(c4.imputed$c4a_expression, c4.imputed$c4b_expression)
cor(c4.imputed$c4a_expression, c4.imputed$c4b_expression)
```

## Count #
```{r}
c4.imputed %>% group_by(c4_allele1_structure) %>% tally()
c4.imputed %>% group_by(c4_allele2_structure) %>% tally()

c4.imputed %>% group_by(c4a_expression) %>% tally()
c4.imputed %>% group_by(c4b_expression) %>% tally()
```

## Save data #
```{r}
c4.imputed <- dplyr::select(c4.imputed, subjectkey, c4a_total, c4b_total, c4herv_total, c4a_dosage, c4b_dosage, avg_dosage, c4_allele1_structure, c4_allele2_structure, c4a_expression, c4b_expression)
write.table(c4.imputed, "../data/ABCD.C4.imputed.multiethnic.txt", sep = "\t", row.names = FALSE, col.names = TRUE, quote = FALSE)
```
